main

+++ 0x03-debugging/main.h
@@ -1,9 +1,10 @@
-#ifndef MAIN_H
+#ifndef MAIN_H
 #define MAIN_H
 
 #include <stdio.h>
+
 void positive_or_negative(int i);
-int largest_number(int a, int b, int c);
+long largest_number(long a, long b, long c);
 void print_remaining_days(int month, int day, int year);
 int convert_day(int month, int day);
-#endif
+#endif /* MAIN_H */
